+++ src/help/help.navigation.tsx
@@ -8,7 +8,7 @@ const Stack = createNativeStackNavigator()
 
 export const HelpNavigation = () => {
   return (
-    <Stack.Navigator screenOptions={{ headerShown: false }}>
+    <Stack.Navigator screenOptions={{ headerShown: false, animation: 'fade', animationDuration: 150 }}>
       <Stack.Screen name={SCREEN.MORE_HELP_MAIN} component={Help} />
       <Stack.Screen name={SCREEN.MORE_HELP_GUIDELINE} component={Guideline} />
     </Stack.Navigator>